+++ core/docs/changelog.md
@@ -1,4 +1,10 @@
 # TBD
+### Breaking Changes
+* Renamed `FilesArtifactID` to `FilesArtifactUUID` which is more accurate
+  * Users should adapt their code to use this new type name when calling `EnclaveContext` methods like `EnclaveContext.StoreWebFiles`
+
+### Changes
+* Replaced `FilesArtifactID` type, provided by `container-engine-lib` with a new type `FilesArtifactUUID`, which is more accurate
 
 # 1.53.5
 ### Changes